% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{clean_tax_table}
\alias{clean_tax_table}
\title{Deal with NAs}
\usage{
clean_tax_table(tax_table)
}
\arguments{
\item{tax_table}{A table obtained by \link[phyloseq]{tax_table}.}
}
\value{
Modified tax table
}
\description{
Replaces NAs with the following: {Class}_NA. Class stands for the last known class of the corresponding OTU.
}
