% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{createAppDir}
\alias{createAppDir}
\title{create App dir

Creates a directory called shiny in \code{location}.
    You can then run the app by just typing \code{runApp("location")}}
\usage{
createAppDir(location)
}
