% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{deseq2_table}
\alias{deseq2_table}
\title{DESeq2 table}
\usage{
deseq2_table(phylo, attribute, conds, formula = NULL, parallel = F)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}

\item{conds}{conditions}
}
\value{
DESeq2 table returned from \link[DESeq2]{results}.
}
\description{
Performs differential expression analysis using DESeq2. If \code{cond1} and \code{cond2} are both NULL then all conditions are taken.
}
