% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{diversity_test}
\alias{diversity_test}
\title{Alpha Diversity test}
\usage{
diversity_test(phylo, attribute)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}
}
\value{
P-value
}
\description{
Perform an anova test on diversity analysis.
}
