% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{generateLineage}
\alias{generateLineage}
\title{Generate Lineage}
\usage{
generateLineage(feature_info)
}
\arguments{
\item{feature_info}{table of features for OTUs}
}
\value{
lineage table
}
\description{
Generates lineage table.
}
