% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{generatePhylo}
\alias{generatePhylo}
\title{Generate Phyloseq object}
\usage{
generatePhylo(study, counts, sample_info, lineage)
}
\arguments{
\item{study}{study ID}

\item{counts}{Table with counts}

\item{sample_info}{Sample_info table}

\item{lineage}{lineage table generated by link{generateLineage}}
}
\value{
phyloseq object
}
\description{
Generates the phyloseq Object of a specific study.
}
