% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{getAttributes}
\alias{getAttributes}
\title{Phyloseq attributes}
\usage{
getAttributes(phylo)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}
}
\value{
character vector of attributes
}
\description{
Get attributes from given phyloseq object.
}
