% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{metatranscriptome methods}
\alias{metatranscriptome methods}
\alias{taxids2names}
\alias{metatranscriptome methods}
\alias{clean_tax_table}
\alias{metatranscriptome methods}
\alias{deseq2_table}
\alias{metatranscriptome methods}
\alias{diversity_test}
\alias{metatranscriptome methods}
\alias{generatePhylo}
\alias{metatranscriptome methods}
\alias{getAttributes}
\alias{metatranscriptome methods}
\alias{loadPhylo}
\alias{metatranscriptome methods}
\alias{generateLineage}
\title{Deal with NAs}
\usage{
taxids2names(phylo, TaxID)

clean_tax_table(tax_table)

deseq2_table(phylo, attribute, cond1 = NULL, cond2 = NULL)

diversity_test(phylo, attribute)

generatePhylo(study, counts, sample_info, lineage)

getAttributes(phylo)

loadPhylo(study)

generateLineage(feature_info)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{tax_table}{A table obtained by \link[phyloseq]{tax_table}.}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}

\item{cond1}{first condition.}

\item{cond2}{second condition.}

\item{study}{study ID}

\item{counts}{Table with counts}

\item{sample_info}{Sample_info table}

\item{lineage}{lineage table generated by link{generateLineage}}

\item{feature_info}{table of features for OTUs}

\item{taxID}{}

\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}

\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{study}{study ID}

\item{envir}{enviroment to load phyloseq object}
}
\value{
Species name

Modified tax table

DESeq2 table returned from \link[DESeq2]{results}.

P-value

phyloseq object

character vector of attributes

lineage table
}
\description{
Replaces NAs with the following: {Class}_NA. Class stands for the last known class of the corresponding OTU.

Performs differential expression analysis using DESeq2. If \code{cond1} and \code{cond2} are both NULL then all conditions are taken.

Perform an anova test on diversity analysis.

Generates the phyloseq Object of a specific study.

Get attributes from given phyloseq object.

Load phyloseq object into memory. Run first phyloseq_generator.r.

Generates lineage table.
}
