% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{metatranscriptome plots}
\alias{metatranscriptome plots}
\alias{plot_top_species}
\alias{metatranscriptome plots}
\alias{plot_volcano}
\alias{metatranscriptome plots}
\alias{plot_sankey}
\alias{metatranscriptome plots}
\alias{plot_taxa}
\alias{metatranscriptome plots}
\alias{plot_diff}
\alias{metatranscriptome plots}
\alias{plot_mds}
\alias{metatranscriptome plots}
\alias{plot_alpha}
\title{Top relative abundance}
\usage{
plot_top_species(phylo, top_n = 10L, attribute)

plot_volcano(tax_table, de_table)

plot_sankey(phylo, source = "Kingdom", target = "Phylum",
  level_filter = NULL, source_filter = NULL)

plot_taxa(phylo, attribute, level, relative = T)

plot_diff(phylo, taxids, attribute)

plot_mds(phylo, color)

plot_alpha(phylo, color)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{top_n}{integer indicating how many species to show}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}

\item{tax_table}{A table obtained by \link[phyloseq]{tax_table}.}

\item{de_table}{A de_table obtained from \link{deseq2_table}.}

\item{source}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")}}

\item{target}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")}}

\item{level_filter}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")}}

\item{source_filter}{String to filter \code{source} values.}

\item{level}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species".)}}

\item{relative}{logical value.}

\item{taxids}{a list of taxids.}

\item{color}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}. The distance methods used is JSD.}

\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}

\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}

\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{color}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}. The distance methods used is JSD.}
}
\value{
A bar plot grouped by \code{attribute}.

An intercative volcano plot generated by plotly.

An intercative sankey generated by plotly.

A barplot grouped by \code{attribute}.

A ggplot2 object showing the boxplots.

A ggplot2 object showing the MDS plot

A ggplot2 object showing the alpha diversity plot.
}
\description{
Plot the relative abundance of the \code{top_n} most abundant species.

Plot the log2(Fold change) against the p-value.

Plot a sankey diagram.

Plot a Taxonomy Bar Chart.

Plot Boxplots for the abundance distribution of the \code{taxids} grouped by \code{attribute}.

Plot MDS of the given phyloseq object, colored by \code{color}.

Plot Alpha diversity of the given phyloseq object, colored by \code{color}.
}
