% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{mfMeans}
\alias{mfMeans}
\title{Metafeature mean relative abundance table}
\usage{
mfMeans(study_info, studies, dir)
}
\arguments{
\item{study_info}{The table study_info}

\item{studies}{A vector of study IDs}

\item{dir}{The data directory}
}
\description{
Create a table of the mean relative abundance of each metafeature per study.
}
