% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_alpha}
\alias{plot_alpha}
\title{Alpha diversity plot}
\usage{
plot_alpha(phylo, color)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{color}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}. The distance methods used is JSD.}
}
\value{
A ggplot2 object showing the alpha diversity plot.
}
\description{
Plot Alpha diversity of the given phyloseq object, colored by \code{color}.
}
