% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_diff}
\alias{plot_diff}
\title{DE Boxplot}
\usage{
plot_diff(phylo, taxids, attribute)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{taxids}{a list of taxids.}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}
}
\value{
A ggplot2 object showing the boxplots.
}
\description{
Plot Boxplots for the abundance distribution of the \code{taxids} grouped by \code{attribute}.
}
