% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_mds}
\alias{plot_mds}
\title{MDS plot}
\usage{
plot_mds(phylo, color)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{color}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}. The distance methods used is JSD.}
}
\value{
A ggplot2 object showing the MDS plot
}
\description{
Plot MDS of the given phyloseq object, colored by \code{color}.
}
