% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_sankey}
\alias{plot_sankey}
\title{Sankey plot}
\usage{
plot_sankey(phylo, source = "Kingdom", target = "Phylum",
  level_filter = NULL, source_filter = NULL)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{source}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")}}

\item{target}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")}}

\item{level_filter}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")}}

\item{source_filter}{String to filter \code{source} values.}
}
\value{
An intercative sankey plot generated by plotly.
}
\description{
Plot a sankey diagram.
}
