% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_taxa}
\alias{plot_taxa}
\title{Taxonomy Bar Chart}
\usage{
plot_taxa(phylo, attribute, level, relative = T)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object.}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}

\item{level}{\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species".)}}

\item{relative}{logical value.}
}
\value{
A barplot grouped by \code{attribute}.
}
\description{
Plot a Taxonomy Bar Chart.
}
