% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_top_species}
\alias{plot_top_species}
\title{Top relative abundance}
\usage{
plot_top_species(phylo, top_n = 10L, attribute)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{top_n}{integer indicating how many species to show}

\item{attribute}{Column name of the sample table, obtained from \link[phyloseq]{sam_data}.}
}
\value{
A bar plot grouped by \code{attribute}.
}
\description{
Plot the relative abundance of the \code{top_n} most abundant species.
}
