% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(tax_table, de_table)
}
\arguments{
\item{tax_table}{A table obtained by \link[phyloseq]{tax_table}.}

\item{de_table}{A de_table obtained from \link{deseq2_table}.}
}
\value{
An intercative volcano plot generated by plotly.
}
\description{
Plot the log2(Fold change) against the p-value.
}
