% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{relativeCounts}
\alias{relativeCounts}
\title{Relative Counts}
\usage{
relativeCounts(phylo)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}
}
\value{
A data frame of relative counts
}
\description{
Divides the counts of each samples by the total number of reads (including non-microbial reads) in each sample.
}
