% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{taxids2names}
\alias{taxids2names}
\title{TaxID to Species

Transforms taxID to the corresponding Species name.}
\usage{
taxids2names(phylo, TaxID)
}
\arguments{
\item{phylo}{\link[phyloseq]{phyloseq} object}

\item{taxID}{}
}
\value{
Species name
}
