% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.r
\name{transformData}
\alias{transformData}
\title{Transform Data

Convert tables to input compatible with the webtool.}
\usage{
transformData(tables, output_dir = pkg_file("data"), log = F)
}
\arguments{
\item{tables}{A directory containing the tables or or list that contains the following files: study_info, sample_info, feature_info, counts}

\item{output_dir}{The directory where the transformed data will be saved. The default is inside the package.}

\item{log}{A logical value indicating whether a log file should be generated with info about the process. The default is \code{FALSE}}
}
