% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{loadPhylo}
\alias{loadPhylo}
\title{Load phyloseq}
\usage{
loadPhylo(study, dir = pkg_file("data"), envir = environment(loadPhylo))
}
\arguments{
\item{study}{study ID}

\item{dir}{directory of the data.}

\item{envir}{enviroment to load phyloseq object}
}
\description{
Load phyloseq object into memory. Run first phyloseq_generator.r.
}
